# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Shopping App - CRUD",
    "description": "My Shopping App Project Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:4000",
  "basePath": "/",
...
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Get all products",
        "parameters": [
          {
            "name": "productName",
            "in": "query",
            "required": false,
            "description": "Product Name",
            "type": "string"
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "description": "Category ID",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ServiceResponse"
            }
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Create Product API",
        "parameters": [
          {
            "name": "productName",
            "in": "formData",
            "description": "Product Name",
            "required": true,
            "type": "string"
          },
          {
            "name": "category",
            "in": "formData",
            "description": "Category ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "productShortDescription",
            "in": "formData",
            "description": "Product Short Description",
            "required": true,
            "type": "string"
          },
          {
            "name": "productDescription",
            "in": "formData",
            "description": "Product Description",
            "type": "string"
          },
          {
            "name": "productPrice",
            "in": "formData",
            "description": "Product Price",
            "required": true,
            "type": "integer"
          },
          {
            "name": "productSalePrice",
            "in": "formData",
            "description": "Product Sale Price",
            "required": true,
            "type": "integer"
          },
          {
            "name": "productImage",
            "in": "formData",
            "description": "Product Image",
            "required": true,
            "type": "file"
          },
          {
            "name": "productSKU",
            "in": "formData",
            "description": "Product SKU",
            "required": false,
            "type": "string"
          },
          {
            "name": "productType",
            "in": "formData",
            "description": "Product Type",
            "required": true,
            "type": "string",
            "default": "simple"
          },
          {
            "name": "Stock Status",
            "in": "formData",
            "description": "Stock Status",
            "required": false,
            "type": "string",
            "default": "IN"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ServiceResponse"
            }
          }
        }
      }
    },
    "/api/products/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Product Id",
          "type": "string"
        }
      ],
      "get": {
        "tags": ["Products"],
        "summary": "Product Details based on Product Id",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ServiceResponse"
            }
          }
        }
      },
      "put": {
        "tags": ["Products"],
        "summary": "Update Product API",
        "parameters": [
          {
            "name": "productName",
            "in": "formData",
            "description": "Product Name",
            "type": "string"
          },
          {
            "name": "category",
            "in": "formData",
            "description": "Category ID",
            "type": "string"
          },
          {
            "name": "productShortDescription",
            "in": "formData",
            "description": "Product Short Description",
            "type": "string"
          },
          {
            "name": "productDescription",
            "in": "formData",
            "description": "Product Description",
            "type": "string"
          },
          {
            "name": "productPrice",
            "in": "formData",
            "description": "Product Price",
            "type": "integer"
          },
          {
            "name": "productSalePrice",
            "in": "formData",
            "description": "Product Sale Price",
            "type": "integer"
          },
          {
            "name": "productImage",
            "in": "formData",
            "description": "Product Image",
            "type": "file"
          },
          {
            "name": "productSKU",
            "in": "formData",
            "description": "Product SKU",
            "type": "string"
          },
          {
            "name": "productType",
            "in": "formData",
            "description": "Product Type",
            "type": "string",
            "default": "simple"
          },
          {
            "name": "Stock Status",
            "in": "formData",
            "description": "Stock Status",
            "type": "string",
            "default": "IN"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ServiceResponse"
            }
          }
        }
      },
      "delete": {
        "tags": ["Products"],
        "summary": "Delete Product API",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ServiceResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Product": {
      "required": ["productName", "_id", "productPrice"],
      "properties": {
        "_id": {
          "type": "integer",
          "uniqueItems": true
        },
        "productName": {
          "type": "string"
        },
        "productDescription": {
          "type": "string"
        },
        "productPrice": {
          "type": "number"
        },
        "productImage": {
          "type": "string"
        }
      }
    },
...
    "ServiceResponse": {
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product"
          }
        }
      }
    }
  }
}